# lib/api-zod/src/generated/types/allocationWithStudent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RealLib - Library Seat Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Student } from './student';

export interface AllocationWithStudent {
  id: number;
  seatId: number;
  studentId: number;
  startTime: string;
  endTime: string;
  isActive: boolean;
  student?: Student;
}