# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RealLib - Library Seat Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  name: string;
  rollNumber: string;
  phoneNumber: string;
  pin: string;
  isActive?: boolean;
  isPaid?: boolean;
}