# lib/api-zod/src/generated/types/unpaidStudent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RealLib - Library Seat Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface UnpaidStudent {
  id: number;
  name: string;
  rollNumber: string;
  phoneNumber: string;
  isActive: boolean;
  isPaid: boolean;
  /** @nullable */
  unpaidSince?: string | null;
  daysUnpaid: number;
  isOverdue: boolean;
  createdAt: string;
}